+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added `Beaker.experiment.rename()`.
+- Added `Beaker.experiment.tasks()`.
 - Added `Beaker.experiment.results()`.
 
 ## [v0.5.5](https://github.com/allenai/beaker-py/releases/tag/v0.5.5) - 2022-04-10