+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `Beaker.experiment.results()`.
 - Added `Beaker.experiment.spec()`.
 - Added `Beaker.workspace.datasets()`.
+- Added `Beaker.workspace.experiments()`.
 
 ## [v0.5.5](https://github.com/allenai/beaker-py/releases/tag/v0.5.5) - 2022-04-10